backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx[http2]==0.27.0
beautifulsoup4==4.12.3
playwright==1.47.0
sse-starlette==2.1.3
